agents/crypto_trader_bot/pyproject.toml
[tool.poetry]
name = "ai-crypto-trader"
version = "0.1.0"
description = ""
authors = ["Tomisin Jenrola <13075885+ToJen@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
python-telegram-bot = "21.7"
aiohttp = "^3.11.8"
swarmzero = "^0.0.3"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
